+++ DistFiles/ReleaseNotes.md
@@ -1,4 +1,12 @@
-﻿## 1.5.2 27 October 2020
+﻿## 1.5.3 8 January 2021
+* Added Help system
+* Added (optional) Edit column
+* Various improvements to filtering and modifying questions.
+
+## 1.5.2 27 October 2020
+* Minimum version needed to produce compatible files to import into Scripture Forge.
+
+## 1.5.2 27 October 2020
 * Minimum version needed to produce compatible files to import into Scripture Forge.
 
 ## 1.5.0 23 September 2020